fix(wire): ARC-008(a) rust-client vector Vec<f32> -> Vec<f64>
Only the Rust client narrowed vectors to f32; the server, TS, and Python
clients all carry full JSON-number (f64) precision, so the Rust client was the
sole path that silently dropped precision on a round-trip. f64 matches the
wire and the other three implementations. No wire-byte change (already
serialized via JSON).

+++ rust-client/src/query.rs
@@ -159,7 +159,7 @@ impl TableQuery {
     pub fn vector_search(
         mut self,
         index: &str,
-        vector: Vec<f32>,
+        vector: Vec<f64>,
         limit: u32,
         filter: BTreeMap<String, serde_json::Value>,
     ) -> Self {
